requirements.txt
asgiref==3.4.1
Django==3.2.8
django-countries==7.2.1
django-filter==21.1
django-widget-tweaks==1.4.9
djangorestframework==3.12.4
importlib-metadata==4.8.1
Markdown==3.3.4
Pillow==8.4.0
pkg-resources==0.0.0
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
zipp==3.6.0
